router: secured ssl flag refactored - mved to config.local.neon Supervizor\Utils => Supervizor\DI Supervizor\Auth => Supervizor\Security removed unused SupervizorSettings
