+++ app/Supervizor/DI/Importer/Exceptions.php
@@ -6,7 +6,7 @@
  * and open the template in the editor.
  */
 
-namespace Supervizor\Utils\Importer\Parsers;
+namespace Supervizor\DI\Importer\Parsers;
 
 /**
  * Description of Exceptions
