+++ app/Supervizor/DI/Importer/ImporterExtension.php
@@ -19,27 +19,41 @@
  * MA 02110-1301  USA
  */
 
-namespace Supervizor\Auth;
+namespace Supervizor\DI\Importer;
 
-use Kdyby\Doctrine\EntityManager;
+use Nette\DI\CompilerExtension;
+use Nette\PhpGenerator\ClassType;
 
-class UserRepository
+/**
+ * Description of EncryptorExtension
+ *
+ * @author Adam Schubert <adam.schubert@sg1-game.net>
+ */
+final class ImporterExtension extends CompilerExtension
 {
 
-    /** @var \Kdyby\Doctrine\EntityRepository */
-    private $userRepository;
+    /** @var array */
+    private $defaults = [
+        'target' => null,
+        'imports' => []
+    ];
 
-    public function __construct(EntityManager $entityManager)
+
+    public function loadConfiguration()
     {
-        $this->userRepository = $entityManager->getRepository(User::class);
+        $config = $this->validateConfig($this->defaults);
+        $builder = $this->getContainerBuilder();
+
+        $builder->addDefinition($this->prefix('importer'))
+                ->setClass(Importer::class, ['@cacheStorage', $config['imports'], '@' . $config['target']]);
     }
 
     /**
-     * @return \Kdyby\Doctrine\EntityRepository
+     * @param ClassType $class
      */
-    public function getUserRepository()
+    public function afterCompile(ClassType $class)
     {
-        return $this->userRepository;
+        
     }
 
 }
