+++ app/Supervizor/Security/Controls/SignInFormFactory.php
@@ -19,7 +19,7 @@
  * MA 02110-1301  USA
  */
 
-namespace Supervizor\Auth\Controls;
+namespace Supervizor\Security\Controls;
 
 interface SignInFormFactory
 {
