+++ app/config/extension.importer.neon
@@ -1,5 +1,5 @@
 extensions:
-  importer: Supervizor\Utils\Importer\ImporterExtension
+  importer: Supervizor\DI\Importer\ImporterExtension
 
 importer:
   target: Supervizor\Import\ImportTarget
@@ -13,5 +13,5 @@ importer:
           description: "Přehled faktur Ministerstva financí"
           source: "http://data.mfcr.cz/cs/api/3/action/resource_show?id=aec18a6a-0d8f-49a4-a8e7-ae0fbd32125f"
           homepage: "http://data.mfcr.cz/cs/dataset/prehled-faktur-ministerstva-financi-cr"
-          parser: Supervizor\Utils\Importer\Parsers\Mfcr
+          parser: Supervizor\DI\Importer\Parsers\Mfcr
           default: true
